refactor: 💡 move registerCustomClaimType to Claims
